requirements.txt
tensorflow
datasets
librosa
numpy
python-dotenv
sounddevice
scipy